requirements.txt
langchain
openai
python-dotenv
python-socketio
gtts
pygame
PyAudio
threading
whisper
SpeechRecognition
google.cloud